backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python==4.8.1.78
numpy==1.24.3
boto3==1.34.34
python-dotenv==1.0.0
